Add enemy archetype management and refactor enemy catalog
- Introduced a new `enemies.ts` module to define enemy archetypes and related functions, enhancing clarity in enemy management.
- Updated the `enemy_catalog.ts` to utilize the new archetype definitions, improving maintainability and consistency in enemy data handling.
- Refactored enemy-related functions to leverage the new archetype structure, streamlining enemy behavior and authoring key retrieval.
- Adjusted various components and prefabs to integrate the new enemy archetype system, ensuring cohesive gameplay mechanics.
- Added tests for the new enemy archetype functionalities to validate correct behavior and error handling.

+++ src/ecs/components.ts
@@ -285,7 +285,7 @@ export const OnTalkEvent: Component<OnTalkEventSchema, typeof ON_TALK_EVENT_STOR
 });
 
 export type TerminalDestinationSchema = { destination: number };
-const TERMINAL_DESTINATION_STORAGE = { destination: Uint16Array };
+const TERMINAL_DESTINATION_STORAGE = { destination: Uint32Array };
 export const TerminalDestination: Component<TerminalDestinationSchema, typeof TERMINAL_DESTINATION_STORAGE> =
   new Component<TerminalDestinationSchema, typeof TERMINAL_DESTINATION_STORAGE>({
     name: "terminalDestination",
